revert worldmap_defs separation
