+++ src/art.cc
@@ -16,7 +16,6 @@
 #include "draw.h"
 #include "game.h"
 #include "memory.h"
-#include "obj_types.h"
 #include "proto.h"
 #include "settings.h"
